refactor logger out
