+++ python/ray/rllib/agents/a3c/a3c_tf_policy_graph.py
@@ -10,6 +10,7 @@ import gym
 import ray
 from ray.rllib.utils.error import UnsupportedSpaceException
 from ray.rllib.utils.explained_variance import explained_variance
+from ray.rllib.utils.timer import setup_custom_logger
 from ray.rllib.evaluation.postprocessing import compute_advantages
 from ray.rllib.evaluation.tf_policy_graph import TFPolicyGraph, \
     LearningRateSchedule
